DZ2 добавлен useEffect
