+++ lesson1/my-app/src/App.js
@@ -1,50 +1,26 @@
-import { Message } from './components'
-import { Button } from './components'
+// import { Message } from './components'
+// import { Button } from './components'
 import React, { useCallback } from 'react'
 import { useState, useEffect } from 'react'
 import { styles } from './App.css'
-import { useCreateMessageForm } from './components/hooks/useCreateMessageForm'
-
-// const onClick = () => {
-//   console.log("клик")
-// }
-
+import { useCreateMessageForm } from './hooks/useCreateMessageForm'
+import { useMessageList } from './hooks/useMessageList'
 
 function App() {
 
-  // const [inputValue, setInputValue] = useState('')
-  const [messageList, setMessageList] = useState([])
-
-  // const onChangeInput = useCallback((event) => {
-  //   setInputValue(event.target.value)
-  // })
-
-  const addNewMessage = useCallback(
-    (message) => {
-      const mes = {
-        author: 'User',
-        message: message
-      }
-      setMessageList((prevState) => {
-        return [
-          ...prevState,
-          mes
-        ]
-      })
-      console.log(messageList)
-    }, [])
-
+  const { messageList, addNewMessage } = useMessageList()
   const { handleSubmit, onChangeInput, inputValue } = useCreateMessageForm({ onSubmit: addNewMessage })
-  // const resetInputValue = useCallback(() => {
-  //   setInputValue('')
-  // }, [])
-
-  // const handleSubmit = useCallback((event) => {
-  //   // отменяем действие по умолчанию
-  //   event.preventDefault()
-  //   addNewMessage(inputValue)
-  //   resetInputValue()
-  // }, [addNewMessage, inputValue, resetInputValue])
+  const messageText = 'Привет'
+  useEffect(() => {
+    if (messageList.length === 0) {
+      return
+    }
+    const text = messageList[messageList.length - 1]
+    if (text.author == 'bot') {
+      return
+    }
+    addNewMessage(messageText, 'bot')
+  }, [messageList])
 
   return (
     <div className="App">
